Partial fix for #2, do not access SOCKOPT_EVENTS so often
Accessing SOCKOPT_EVENTS nullifies the edge-triggered read/write events.
We need to only get the option once and check it against POLL_IN/POLL_OUT,
so that the read event gets triggered, and we do not accidentally one of
the read events, leading to a empty pull socket.

Includes a test script reproducing the problems.

Unfortunately the fix currently causes an infinite loop with the pubsub
example and the push/pull integration test.