Emit an explicit messages event for multi-part receive
The current behaviour is inconsistent because the message event can be
either a single message or in case of a multi-part message, an array.

This patch makes an explicit distinction. The message event is used for
single part messages, the messages even for multi-part. The messages
event is also emitted for single part messages, where the argument is an
array of one single element.

Closes #3.